+++ cate/cli/main.py
@@ -105,7 +105,8 @@ from typing import Tuple, Union, List, Dict, Any, Optional
 
 from cate.conf.defaults import WEBAPI_INFO_FILE, WEBAPI_ON_INACTIVITY_AUTO_STOP_AFTER
 from cate.core.types import Like, TimeRangeLike, PolygonLike, VarNamesLike
-from cate.core.ds import DATA_STORE_REGISTRY, find_data_sources
+from cate.core.ds import DATA_STORE_REGISTRY, find_data_sources, format_cached_datasets_coverage_string, \
+    format_variables_info_string
 from cate.core.objectio import OBJECT_IO_REGISTRY, find_writer, read_object
 from cate.core.op import OP_REGISTRY
 from cate.core.plugin import PLUGIN_REGISTRY
@@ -342,8 +343,12 @@ def _get_op_io_info_str(inputs_or_outputs: dict, title_singular: str, title_plur
     op_info_str = ''
     op_info_str += '\n'
     if inputs_or_outputs:
+        inputs_or_outputs = {name: properties for name, properties in inputs_or_outputs.items()
+                             if not properties.get('deprecated')}
         op_info_str += '%s:' % (title_singular if len(inputs_or_outputs) == 1 else title_plural)
         for name, properties in inputs_or_outputs.items():
+            if properties.get('deprecated'):
+                continue
             op_info_str += '\n'
             op_info_str += '  %s (%s)' % (name, _get_op_data_type_str(properties.get('data_type', object)))
             description = properties.get('description', None)
@@ -1028,6 +1033,8 @@ class OperationCommand(SubCommandCommand):
                                  help="List only operations tagged by TAG or "
                                       "that have TAG in one of their tags. "
                                       "The comparison is case insensitive.")
+        list_parser.add_argument('--deprecated', '-d', action='store_true',
+                                 help="List deprecated operations.")
         list_parser.add_argument('--internal', '-i', action='store_true',
                                  help='List operations tagged "internal".')
         list_parser.set_defaults(sub_command_function=cls._execute_list)
@@ -1041,11 +1048,13 @@ class OperationCommand(SubCommandCommand):
     def _execute_list(cls, command_args):
         op_regs = OP_REGISTRY.op_registrations
 
-        def _is_op_selected(op_reg, tag_part: str, is_internal: bool):
+        def _is_op_selected(op_reg, tag_part: str, internal_only: bool, deprecated_only: bool):
+            if deprecated_only and not op_reg.op_meta_info.header.get('deprecated'):
+                return False
             tags = to_list(op_reg.op_meta_info.header.get('tags'))
             if tags:
                 # Tagged operations
-                if is_internal:
+                if internal_only:
                     if 'internal' not in tags:
                         return False
                 else:
@@ -1057,13 +1066,13 @@ class OperationCommand(SubCommandCommand):
                         return any(tag_part in tag.lower() for tag in tags)
                     elif isinstance(tags, str):
                         return tag_part in tags.lower()
-            elif is_internal or tag_part:
+            elif internal_only or tag_part:
                 # Untagged operations
                 return False
             return True
 
         op_names = sorted([op_name for op_name, op_reg in op_regs.items() if
-                           _is_op_selected(op_reg, command_args.tag, command_args.internal)])
+                           _is_op_selected(op_reg, command_args.tag, command_args.internal, command_args.deprecated)])
         name_pattern = None
         if command_args.name:
             name_pattern = command_args.name
@@ -1184,13 +1193,13 @@ class DataSourceCommand(SubCommandCommand):
             print('\n'
                   'Locally stored datasets:\n'
                   '------------------------\n'
-                  '{info}'.format(info=data_source.cached_datasets_coverage_string))
+                  '{info}'.format(info=format_cached_datasets_coverage_string(data_source.cache_info)))
         if command_args.var:
             print()
             print('Variables')
             print('---------')
             print()
-            print(data_source.variables_info_string)
+            print(format_variables_info_string(data_source.variables_info))
 
     @classmethod
     def _execute_add(cls, command_args):