+++ cate/conf/conf.py
@@ -22,9 +22,10 @@
 __author__ = "Norman Fomferra (Brockmann Consult GmbH)"
 
 import os.path
-from typing import Any
+from typing import Any, Dict, Optional
 
-from .defaults import GLOBAL_CONF_FILE, LOCAL_CONF_FILE, LOCATION_FILE, VERSION_CONF_FILE
+from .defaults import GLOBAL_CONF_FILE, LOCAL_CONF_FILE, LOCATION_FILE, VERSION_CONF_FILE, \
+    VARIABLE_DISPLAY_SETTINGS, DEFAULT_COLOR_MAP
 
 _CONFIG = None
 
@@ -53,6 +54,22 @@ def get_config_value(name: str, default=None) -> Any:
     if not name:
         raise ValueError('name argument must be given')
     return get_config().get(name, default)
+
+
+def get_variable_display_settings(var_name: str) -> Optional[Dict[str, Any]]:
+    """
+    Get the global variable display settings which is a combination of defaults.
+    :return:
+    """
+    settings_dict = get_config().get('variable_display_settings', None)
+    if settings_dict and var_name in settings_dict:
+        return settings_dict[var_name]
+
+    settings = VARIABLE_DISPLAY_SETTINGS.get(var_name)
+    if settings:
+        return settings
+
+    return dict(color_map=get_config_value('default_color_map', DEFAULT_COLOR_MAP))
 
 
 def get_config() -> dict: