+++ cate/ds/esa_cci_odp.py
@@ -59,9 +59,11 @@ from cate.conf import get_config_value
 from cate.conf.defaults import NETCDF_COMPRESSION_LEVEL
 from cate.core.ds import DATA_STORE_REGISTRY, DataStore, DataSource, Schema, \
     open_xarray_dataset, get_data_stores_path
-from cate.core.types import PolygonLike, TimeRange, TimeRangeLike, VarNamesLike, VarNames
+from cate.core.types import PolygonLike, TimeLike, TimeRange, TimeRangeLike, VarNamesLike, VarNames
 from cate.ds.local import add_to_data_store_registry, LocalDataSource, LocalDataStore
 from cate.util.monitor import Monitor
+
+ESA_CCI_ODP_DATA_STORE_ID = 'esa_cci_odp'
 
 __author__ = "Norman Fomferra (Brockmann Consult GmbH), " \
              "Marco Zühlke (Brockmann Consult GmbH), " \
@@ -114,7 +116,7 @@ def get_data_store_path():
 
 def get_metadata_store_path():
     return os.environ.get('CATE_ESA_CCI_ODP_DATA_STORE_PATH',
-                          os.path.join(get_data_stores_path(), 'esa_cci_odp'))
+                          os.path.join(get_data_stores_path(), ESA_CCI_ODP_DATA_STORE_ID))
 
 
 def set_default_data_store():
@@ -495,6 +497,10 @@ class EsaCciOdpDataSource(DataSource):
         return None
 
     @property
+    def variables_info(self):
+        return self._variables_list()
+
+    @property
     def schema(self) -> Schema:
         return self._schema
 
@@ -629,7 +635,8 @@ class EsaCciOdpDataSource(DataSource):
 
         if to_add:
             for time_range_to_add in to_add:
-                self._make_local(data_source, time_range_to_add, None, data_source.variables_info, monitor)
+                self._make_local(data_source, time_range_to_add, None, [var.get('name') for var
+                                                                        in data_source.variables_info], monitor)
             data_source.meta_info['temporal_coverage_start'] = time_range[0]
             data_source.meta_info['temporal_coverage_end'] = time_range[1]
             data_source.update_temporal_coverage(time_range)
@@ -726,7 +733,6 @@ class EsaCciOdpDataSource(DataSource):
                     monitor: Monitor = Monitor.NONE):
 
         local_id = local_ds.id
-        local_ds.meta_info
         time_range = TimeRangeLike.convert(time_range)
         region = PolygonLike.convert(region)
         var_names = VarNamesLike.convert(var_names)
@@ -819,10 +825,10 @@ class EsaCciOdpDataSource(DataSource):
                                 lon_min = geo_lon_max - lon_max
                                 lon_max = geo_lon_max - lon_min_copy
 
-                            lat_min = floor(lat_min / geo_lat_res)
-                            lat_max = ceil(lat_max / geo_lat_res)
-                            lon_min = floor(lon_min / geo_lon_res)
-                            lon_max = ceil(lon_max / geo_lon_res)
+                            lat_min = int(floor(lat_min / geo_lat_res))
+                            lat_max = int(ceil(lat_max / geo_lat_res))
+                            lon_min = int(floor(lon_min / geo_lon_res))
+                            lon_max = int(ceil(lon_max / geo_lon_res))
 
                             remote_dataset = remote_dataset.isel(drop=False,
                                                                  lat=slice(lat_min, lat_max),
@@ -844,14 +850,18 @@ class EsaCciOdpDataSource(DataSource):
                                 geo_lon_max_copy = geo_lon_max
                                 geo_lon_min = geo_lon_max_copy - lon_max * geo_lon_res
                                 geo_lon_max = geo_lon_max_copy - lon_min * geo_lon_res
+                    print(var_names)
                     if not var_names:
                         var_names = [var_name for var_name in remote_netcdf.variables.keys()]
                     var_names.extend([coord_name for coord_name in remote_dataset.coords.keys()
                                       if coord_name not in var_names])
+                    print(var_names)
                     child_monitor.start(label=file_name, total_work=len(var_names))
                     for sel_var_name in var_names:
                         var_dataset = remote_dataset.drop(
                             [var_name for var_name in remote_dataset.variables.keys() if var_name != sel_var_name])
+                        print(remote_dataset)
+                        print(var_dataset)
                         if compression_enabled:
                             var_dataset.variables.get(sel_var_name).encoding.update(encoding_update)
                         local_netcdf.store_dataset(var_dataset)
@@ -924,8 +934,8 @@ class EsaCciOdpDataSource(DataSource):
                             do_update_of_verified_time_coverage_start_once = False
                         verified_time_coverage_end = coverage_to
 
-        local_ds.meta_info['temporal_coverage_start'] = verified_time_coverage_start
-        local_ds.meta_info['temporal_coverage_end'] = verified_time_coverage_end
+        local_ds.meta_info['temporal_coverage_start'] = TimeLike.format(verified_time_coverage_start)
+        local_ds.meta_info['temporal_coverage_end'] = TimeLike.format(verified_time_coverage_end)
         local_ds.save(True)
 
     def _apply_make_local_fixes(self,