+++ cate/webapi/websocket.py
@@ -28,12 +28,12 @@ from typing import List, Sequence
 import xarray as xr
 
 from cate.conf import conf
-from cate.conf.defaults import GLOBAL_CONF_FILE, WEBAPI_USE_WORKSPACE_IMAGERY_CACHE
+from cate.conf.defaults import VERSION_CONF_FILE, WEBAPI_USE_WORKSPACE_IMAGERY_CACHE
 from cate.core.ds import DATA_STORE_REGISTRY, get_data_stores_path, find_data_sources
 from cate.core.op import OP_REGISTRY
 from cate.core.workspace import OpKwArgs
 from cate.core.wsmanag import WorkspaceManager
-from cate.util import Monitor, cwd
+from cate.util import Monitor, cwd, filter_fileset
 
 
 # noinspection PyMethodMayBeStatic
@@ -62,7 +62,7 @@ class WebSocketService:
         # noinspection PyBroadException
         conf_text = ''
         try:
-            with open(GLOBAL_CONF_FILE, 'r') as fp:
+            with open(VERSION_CONF_FILE, 'r') as fp:
                 conf_text = fp.read()
         except:
             # ok
@@ -96,7 +96,7 @@ class WebSocketService:
 
         # Now join lines back again and write modified config file
         conf_text = '\n'.join(conf_lines)
-        with open(GLOBAL_CONF_FILE, 'w') as fp:
+        with open(VERSION_CONF_FILE, 'w') as fp:
             fp.write(conf_text)
 
     def get_data_stores(self) -> list:
@@ -120,7 +120,15 @@ class WebSocketService:
         data_store = DATA_STORE_REGISTRY.get_data_store(data_store_id)
         if data_store is None:
             raise ValueError('Unknown data store: "%s"' % data_store_id)
-        data_sources = sorted(data_store.query(monitor=monitor), key=lambda ds: ds.title or ds.id)
+        data_sources = data_store.query(monitor=monitor)
+        if data_store_id != 'local':
+            data_source_dict = {ds.id: ds for ds in data_sources}
+            data_source_ids = filter_fileset(data_source_dict.keys(),
+                                             includes=conf.get_config_value('included_data_sources', default=None),
+                                             excludes=conf.get_config_value('excluded_data_sources', default=None))
+            data_sources = [data_source_dict[id] for id in data_source_ids]
+
+        data_sources = sorted(data_sources, key=lambda ds: ds.title or ds.id)
         return [dict(id=data_source.id,
                      title=data_source.title,
                      meta_info=data_source.meta_info) for data_source in data_sources]
@@ -193,11 +201,11 @@ class WebSocketService:
                                    monitor=monitor.child(98))
             return self.get_data_sources('local', monitor=monitor.child(2))
 
-    def add_local_datasource(self, data_source_name: str, file_path_pattern: str, monitor: Monitor):
+    def add_local_data_source(self, data_source_id: str, file_path_pattern: str, monitor: Monitor):
         """
-        Adds a local datas ource made up of the specified files.
+        Adds a local data source made up of the specified files.
 
-        :param data_source_name: The name of the local data source.
+        :param data_source_id: The identifier of the local data source.
         :param file_path_pattern: The files path containing wildcards.
         :param monitor: a progress monitor.
         :return: JSON-serializable list of 'local' data sources, sorted by name.
@@ -205,37 +213,44 @@ class WebSocketService:
         data_store = DATA_STORE_REGISTRY.get_data_store('local')
         if data_store is None:
             raise ValueError('Unknown data store: "%s"' % 'local')
-        with monitor.starting('Making data source local', 100):
+        with monitor.starting('Adding local data source', 100):
             # TODO use monitor, while extracting metadata
-            data_store.add_pattern(data_source_id=data_source_name, files=file_path_pattern)
+            data_store.add_pattern(data_source_id=data_source_id, files=file_path_pattern)
             return self.get_data_sources('local', monitor=monitor.child(100))
 
-    def remove_local_datasource(self, data_source_name: str, remove_files: bool) -> list:
+    def remove_local_data_source(self, data_source_id: str, remove_files: bool, monitor: Monitor) -> list:
         """
         Removes the datasource (and optionally the giles belonging  to it) from the local data store.
 
-        :param data_source_name: The name of the local data source.
+        :param data_source_id: The identifier of the local data source.
         :param remove_files: Wether to remove the files belonging to this data source.
+        :param monitor: a progress monitor.
         :return: JSON-serializable list of 'local' data sources, sorted by name.
         """
         data_store = DATA_STORE_REGISTRY.get_data_store('local')
         if data_store is None:
             raise ValueError('Unknown data store: "%s"' % 'local')
-        data_store.remove_data_source(data_source_name, remove_files)
-        return self.get_data_sources('local', monitor=Monitor.NONE)
+        # TODO use monitor, while removing files
+        data_store.remove_data_source(data_source_id, remove_files)
+        return self.get_data_sources('local', monitor=monitor)
 
-    def get_operations(self) -> List[dict]:
+    def get_operations(self, registry=None) -> List[dict]:
         """
         Get registered operations.
 
         :return: JSON-serializable list of data sources, sorted by name.
         """
+        registry = registry or OP_REGISTRY
         op_list = []
-        for op_name, op_reg in OP_REGISTRY.op_registrations.items():
+        for op_name, op_reg in registry.op_registrations.items():
+            if op_reg.op_meta_info.header.get('deprecated'):
+                continue
             op_json_dict = op_reg.op_meta_info.to_json_dict()
             op_json_dict['name'] = op_name
-            op_json_dict['inputs'] = [dict(name=name, **props) for name, props in op_json_dict['inputs'].items()]
-            op_json_dict['outputs'] = [dict(name=name, **props) for name, props in op_json_dict['outputs'].items()]
+            op_json_dict['inputs'] = [dict(name=name, **props) for name, props in op_json_dict['inputs'].items()
+                                      if not props.get('deprecated')]
+            op_json_dict['outputs'] = [dict(name=name, **props) for name, props in op_json_dict['outputs'].items()
+                                       if not props.get('deprecated')]
             op_list.append(op_json_dict)
 
         return sorted(op_list, key=lambda op: op['name'])
@@ -331,7 +346,9 @@ class WebSocketService:
         from cate.util.im.cmaps import get_cmaps
         return get_cmaps()
 
-    def get_workspace_variable_statistics(self, base_dir: str, res_name: str, var_name: str, var_index: Sequence[int]):
+    # Note, we should turn this into an operation "actual_min_max(ds, var)"
+    def get_workspace_variable_statistics(self, base_dir: str, res_name: str, var_name: str, var_index: Sequence[int],
+                                          monitor=Monitor.NONE):
         workspace_manager = self.workspace_manager
         workspace = workspace_manager.get_workspace(base_dir)
         if res_name not in workspace.resource_cache:
@@ -348,7 +365,10 @@ class WebSocketService:
         if var_index:
             variable = variable[tuple(var_index)]
 
-        valid_min = variable.min(skipna=True)
-        valid_max = variable.max(skipna=True)
+        with monitor.starting('Computing min/max', total_work=100.):
+            with monitor.child(work=50.).observing('Computing min'):
+                actual_min = variable.min(skipna=True)
+            with monitor.child(work=50.).observing('Computing max'):
+                actual_max = variable.max(skipna=True)
 
-        return dict(min=float(valid_min), max=float(valid_max))
+        return dict(min=float(actual_min), max=float(actual_max))